+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * Default positionInParent on MPCommandFactory.addElements() changed to mpAddChildAtEndMinusOneOfParentChildrenList.
   * Using a cache for updating THIDs of duplicated elements and their options.
   * Testing line duplication.
+  * Created MPTHElementDuplicatorAux class.
 
 ## 0.3.0 - 2026-03-09 - The [Great Fanfarra](https://youtu.be/HlV59UbrY-E) release
 * New features: